Tweak handling of gear
Check for merging of quiver stacks in combine_pack().  Resolves https://github.com/NickMcConnell/NarSil/issues/296 .  Move weight limit check from inven_carry_okay() to inven_carry_num() so it can be imposed when picking up more than one object from a stack.   Update test in inven_carry_num() for how the number that can go to the quiver is calculated.  When wielding ammunition, prefer to merge with an already occupied quiver slot and enforce the weight limit.  In do_cmd_wield(), only combine the pack if picking up from the floor to avoid the possibility of a dangling reference if the object taken off combines with the object to be wielded.