Smithing: fix handling of special bonus
Have ego_apply_magic() match Sil 1.3's behavior for the special bonus when the item is cursed or the special bonus does not matter.  When including the special bonus during smithing, replace the existing bonus regardless of sign.  Set the starting special bonus for a newly selected kind of object, new enchantment, or new artifact to match the default for that kind and enchantment.  Include the special bonus when deciding whether or not to enable the "Accept" entry.

Resolves https://github.com/NickMcConnell/NarSil/issues/357 .  Resolves https://github.com/NickMcConnell/NarSil/issues/397 .