SDL, SDL2, Mac: port Vanilla Angband's changes limiting tileset changes
For all, disable changing the tile set and tile multipliers when in an in-game menu.  For SDL and SDL2, also disable those changes until the character is fully in game; Mac does allow changes while on the splash screen.

For the Mac, restructure the interface setting the tile multipliers.  Use a single factor relative to the native size rather than setting the tile multipliers directly.  The recompiled Mac interface file was generated with Xcode 14.2.