Refine assignments for tile sets from Vanilla Angband
Fixes parsing errors and chooses close matches for things are present in NarSil.  Allow expressions in the preferences that select on the player's sex since the Shockbolt tiles have different gendered representations for the player's character.  There are still many missing assignments so playing with the tile sets from Vanilla Angband is not yet a smooth experience.