Restructure mithril melting to tolerate weights greater than 255
Has no effect in an unmodded game since the mithril corslet, with a weight of 150, is the heaviest mithril item.  Resolves the fifth part of https://github.com/NickMcConnell/NarSil/issues/735 .  Also correct the calculation of the slots needed for melting when the item's weight is a multiple of mithril's maximum stack size, 99.