In the tutorial, set up known version of starting kit as normal birth does
Also plug memory leaks for the known versions of objects if the tutorial purges the default starting kit or can not fit an item from the starting kit in the pack.  Changes have no effect on the unmodded game as the tutorial currently does not modify the default starting kit for the player.